pyproject.toml
[project]
name = "saimid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-ninja>=1.4.0",
    "django-simple-history>=3.8.0",
    "openai>=1.69.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "nitpick",
 "django-stubs-ext>=5.1.1",
 "django-stubs[compatible-mypy]>=5.1.1",
 "pylint-django>=2.6.1",
 "pytest-django>=4.9.0",
 "basedpyright>=1.21.0",
 "beartype>=0.19.0",
 "mypy>=1.13.0",
 "pyre-check>=0.9.23",
 "pyright>=1.1.388",
 "pytype>=2024.10.11",
 "typeguard>=4.4.1",
 "pylint>=3.3.4",
 "import-linter>=2.1",
 "isort>=5.13.2",
 "ruff>=0.7.3",
 "semgrep>=1.85.0",
 "wemake-python-styleguide>=1.0.0",
 "taplo>=0.9.3",
 "project-config>=0.9.7",
 "nox>=2024.10.9",
 "pre-commit>=4.1.0",
 "python-semantic-release>=9.14.0",
 "hypothesis>=6.118.0",
 "pytest-cov>=6.0.0",
 "pytest-instafail>=0.5.0",
 "pytest-testmon>=2.1.1",
 "pytest-watch>=4.2.0",
 "pytest>=8.3.3",
 "types-requests>=2.32.0.20250328",
]

[tool.uv.sources]
nitpick = { git = "https://github.com/NextGenContributions/nitpick", rev = "develop-ng" }

[tool.semantic_release]
assets = []
build_command_env = []
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "angular"
logging_use_named_masks = false
major_on_zero = true
allow_zero_version = true
no_git_verify = false
tag_format = "v{version}"
version_toml = ["pyproject.toml:project.version"]

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
exclude_commit_patterns = ["chore(?:\\([^)]*?\\))?: .+", "ci(?:\\([^)]*?\\))?: .+", "refactor(?:\\([^)]*?\\))?: .+", "style(?:\\([^)]*?\\))?: .+", "test(?:\\([^)]*?\\))?: .+", "build\\((?!deps\\): .+)", "Merged? .*", "Initial Commit.*", "^\\d+\\.\\d+\\.\\d+"]
mode = "init"
insertion_flag = "<!-- version list -->"
template_dir = "templates"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"
mask_initial_release = false

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = false

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
allowed_tags = ["feat", "fix", "perf", "build", "chore", "ci", "docs", "style", "refactor", "test"]
default_bump_level = 0

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true
